refactor: Post,PostEdit,PostWrite 페이지를 하나로 묶음
