+++ breaking-front/src/App.js
@@ -14,12 +14,12 @@ import SignUp from 'pages/Authentication/SignUp/SignUp';
 import Profile from 'pages/Profile/Profile/Profile';
 import ProfileEdit from 'pages/Profile/ProfileEdit/ProfileEdit';
 import MainFeed from 'pages/MainFeed/MainFeed';
-import Post from 'pages/Post/Post';
-import PostWrite from 'pages/PostWrite/PostWrite';
+import Post from 'pages/Post/Post/Post';
+import PostWrite from 'pages/Post/PostWrite/PostWrite';
 import Financial from 'pages/Financial/Financial';
 import theme from 'styles/theme';
 import GlobalStyle from 'styles/GlobalStyle';
-import PostEdit from 'pages/PostEdit/PostEdit';
+import PostEdit from 'pages/Post/PostEdit/PostEdit';
 import SearchUnified from 'pages/Search/SearchUnified/SearchUnified';
 import SearchPost from 'pages/Search/SearchPost/SearchPost';
 import SearchHashtag from 'pages/Search/SearchHashtag/SearchHashtag';
