+++ breaking-front/src/pages/Post/Post/Post.js
@@ -12,15 +12,15 @@ import ScrollToTop from 'components/ScrollToTop/ScrollToTop';
 import ContentHashtag from 'components/ContentHashtag/ContentHashtag';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
 import usePost from 'hooks/queries/usePost';
-import usePostLike from 'pages/Post/hooks/mutations/usePostLike';
-import useDeletePostLike from 'pages/Post/hooks/mutations/useDeletePostLike';
-import usePostLikeList from 'pages/Post/hooks/queries/usePostLikeList';
-import Carousel from 'pages/Post/components/Carousel/Carousel';
-import CommentList from 'pages/Post/components/CommentList/CommentList';
-import ContentToggle from 'pages/Post/components/ContentToggle/ContentToggle';
-import PurchaseButton from 'pages/Post/components/PurchaseButton/PurchaseButton';
-import PostLikeListModal from 'pages/Post/components/PostLikeListModal/PostLikeListModal';
-import * as Style from 'pages/Post/Post.styles';
+import usePostLike from 'pages/Post/Post/hooks/mutations/usePostLike';
+import useDeletePostLike from 'pages/Post/Post/hooks/mutations/useDeletePostLike';
+import usePostLikeList from 'pages/Post/Post/hooks/queries/usePostLikeList';
+import Carousel from 'pages/Post/Post/components/Carousel/Carousel';
+import CommentList from 'pages/Post/Post/components/CommentList/CommentList';
+import ContentToggle from 'pages/Post/Post/components/ContentToggle/ContentToggle';
+import PurchaseButton from 'pages/Post/Post/components/PurchaseButton/PurchaseButton';
+import PostLikeListModal from 'pages/Post/Post/components/PostLikeListModal/PostLikeListModal';
+import * as Style from 'pages/Post/Post/Post.styles';
 import { ReactComponent as LocationIcon } from 'assets/svg/location.svg';
 import { ReactComponent as LikeIcon } from 'assets/svg/like.svg';
 import { ReactComponent as LikedIcon } from 'assets/svg/liked.svg';
