+++ breaking-front/src/pages/Post/Post/components/Comment/Comment.js
@@ -7,15 +7,15 @@ import timeFormatter from 'utils/timeFormatter';
 import ImageUrlConverter from 'utils/ImageUrlConverter';
 import ContentHashtag from 'components/ContentHashtag/ContentHashtag';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
-import useCommentLike from 'pages/Post/hooks/mutations/useCommentLike';
-import useDeleteCommentLike from 'pages/Post/hooks/mutations/useDeleteCommentLike';
-import useDeleteComment from 'pages/Post/hooks/mutations/useDeleteComment';
-import useCommentReplyWrite from 'pages/Post/hooks/mutations/useCommentReplyWrite';
-import useCommentEdit from 'pages/Post/hooks/mutations/useCommentEdit';
-import Reply from 'pages/Post/components/Reply/Reply';
-import CommentForm from 'pages/Post/components/CommentForm/CommentForm';
-import CommentToggle from 'pages/Post/components/CommentToggle/CommentToggle';
-import * as Style from 'pages/Post/components/Comment/Comment.styles';
+import useCommentLike from 'pages/Post/Post/hooks/mutations/useCommentLike';
+import useDeleteCommentLike from 'pages/Post/Post/hooks/mutations/useDeleteCommentLike';
+import useDeleteComment from 'pages/Post/Post/hooks/mutations/useDeleteComment';
+import useCommentReplyWrite from 'pages/Post/Post/hooks/mutations/useCommentReplyWrite';
+import useCommentEdit from 'pages/Post/Post/hooks/mutations/useCommentEdit';
+import Reply from 'pages/Post/Post/components/Reply/Reply';
+import CommentForm from 'pages/Post/Post/components/CommentForm/CommentForm';
+import CommentToggle from 'pages/Post/Post/components/CommentToggle/CommentToggle';
+import * as Style from 'pages/Post/Post/components/Comment/Comment.styles';
 import { ReactComponent as LikeIcon } from 'assets/svg/like.svg';
 import { ReactComponent as LikedIcon } from 'assets/svg/liked.svg';
 import { ReactComponent as ETCIcon } from 'assets/svg/etc.svg';
