+++ breaking-front/src/pages/Post/Post/components/ContentToggle/ContentToggle.js
@@ -3,14 +3,14 @@ import { useNavigate } from 'react-router-dom';
 import { PAGE_PATH } from 'constants/path';
 import PropTypes from 'prop-types';
 import { UserInformationContext } from 'providers/UserInformationProvider';
-import useDeletePost from 'pages/Post/hooks/mutations/useDeletePost';
+import useDeletePost from 'pages/Post/Post/hooks/mutations/useDeletePost';
 import usePostBookmark from 'hooks/mutations/usePostBookmark';
 import useDeletePostBookmark from 'hooks/mutations/useDeletePostBookmark';
-import usePostActivatePurchase from 'pages/Post/hooks/mutations/usePostActivatePurchase';
-import usePostDeactivatePurchase from 'pages/Post/hooks/mutations/usePostDeactivatePurchase';
+import usePostActivatePurchase from 'pages/Post/Post/hooks/mutations/usePostActivatePurchase';
+import usePostDeactivatePurchase from 'pages/Post/Post/hooks/mutations/usePostDeactivatePurchase';
 import Toggle from 'components/Toggle/Toggle';
 import ShareModal from 'components/ShareModal/ShareModal';
-import * as Style from 'pages/Post/components/ContentToggle/ContentToggle.styles';
+import * as Style from 'pages/Post/Post/components/ContentToggle/ContentToggle.styles';
 import { ReactComponent as EditIcon } from 'assets/svg/edit.svg';
 import { ReactComponent as RemoveIcon } from 'assets/svg/remove.svg';
 import { ReactComponent as ActivateIcon } from 'assets/svg/activate_purchase.svg';
