+++ breaking-front/src/pages/Post/Post/components/ContentToggle/ContentToggle.styles.js
