+++ breaking-front/src/pages/Post/Post/components/PostBoughtListModal/PostBoughtListModal.styles.js
